自定义字段初步调试

+++ qinjee-parent/qinjee-masterdata/src/main/java/com/qinjee/masterdata/model/entity/UserArchive.java
@@ -132,7 +132,7 @@ public class UserArchive implements Serializable {
      */
     @NotNull
     @ApiModelProperty("年龄")
-    private Short age;
+    private Integer age;
 
     /**
      * 民族
